Comment out tutorial modifier
It seems a lot of people would rather drag and drop this plugin in their server hoping it will solve all their problems without taking the time to configure it. In order to reduce the number of people coming to me with the exact same problem I will leave the tutorial modifier commented out so people are still able to read as they wish.